55: extract `<NavigationBar>` and its test into separate files
